Move delete account button to main preferences and style
