+++ react-app/src/components/styles/navbar.css
@@ -400,8 +400,6 @@ font-family: 'Sofia', cursive; */
     color: #33A9B3;
 }
 
-
-
 .nav-links-right {
     height: 140%;
     display: flex;
