+++ react-app/src/components/styles/show-preferences.css
@@ -8,7 +8,7 @@ font-family: 'Sofia', cursive; */
     background-color: #F6F1F6;
     margin-top: -400px;
     margin-bottom: -20px;
-    padding: 25rem 25rem 2rem 25rem;
+    padding: 25rem 25rem 43rem 25rem;
     display: flex;
     flex-direction: column;
     justify-content: center;
@@ -77,7 +77,7 @@ font-family: 'Sofia', cursive; */
 
 .dark-purple-container {
     background-color: #D5BED5;
-    width: 100%;
+    width: 60vw;
     border-radius: 4px;
     margin: .5rem .4rem .3rem -2rem;
 }
@@ -85,7 +85,7 @@ font-family: 'Sofia', cursive; */
 .light-purple-container {
     border-radius: 4px;
     background-color: #E7DAE7;
-    width: 100%;
+    width: 60vw;
     border-radius: 4px;
     margin: .25rem .4rem .25rem -2rem;
 }
@@ -129,6 +129,26 @@ font-family: 'Sofia', cursive; */
 }
 
 #show-preferences-subscription-btn:hover {
+    border: 2px solid #8B478B;
+    background-color: #8B478B;
+}
+
+#delete-account-btn {
+    position: absolute;
+    margin: 4rem .5rem 0 45.5rem;
+    border-radius: 28px;
+    border: 2px solid #420F42;
+    background-color: #420F42;
+    color: white;
+    width: 20rem;
+    height: 3rem;
+    font-family: 'Montserrat', sans-serif;
+    font-size: 16px;
+    margin-bottom: 1rem;
+    cursor: pointer;
+}
+
+#delete-account-btn:hover {
     border: 2px solid #8B478B;
     background-color: #8B478B;
 }
